Unify styling into one primitive (remove the style_/css split)
Description: We had shipped two styling builders — style_ (inline style attribute) and css() (classes + extracted stylesheet). They are not two kinds of thing: both are CSS, both are style. The only real difference is where the declarations land (inline attribute vs stylesheet rule), and the sole capability that buys is selectors (:hover/:focus/media) which inline cannot express. Naming one 'style' and the other 'css' implies a category distinction that does not exist — a misuse of established terms and unnecessary surface. This collapses to one primitive.

Changes: Renamed the class/sheet-based builder css -> style_ (Style/usecase/css.ts -> style_.ts) and deleted the old inline style_ (and its spec). The one style_ now does everything: a hook string + typed atoms + hover/focus/active variants -> content-hashed atomic classes, with collectCss inlining the exact sheet into <head>. The internal Box tag 'Css' and matcher css$ are kept (renaming would churn all seven exhaustive match arms for no user-facing gain). The dynamic-inline case uses the existing attr('style', ...) escape, now documented. Migrated the example: folded every class_ styling hook into a style_ string arg (style_ is the sole class authority, so class_ + style_ would write class twice); class_ remains only for the one non-styling hook (header). Updated specs (css->style_, the renderer's inline-style assertion became a class assertion) and the example README (dropped the 'two modes' framing). Redeployed the dev image + container.

Test Planning: plgg-view tsc clean + 94 vitest tests (was 99; removed the now-redundant inline-style tests); coverage 98.4% statements / 96.1% branches / 97.1% functions / 98.3% lines (gate 91%). plgg-server tsc + 75 tests; example tsc + 17 tests (class hooks preserved via style_ string args, so .todo/.filter/.selected selectors still resolve). core plgg tsc clean. Verified the redeployed container: curl localhost:3001/ -> 200, <head> <style> present with :hover and :focus rules, body elements carry the class hooks, and no inline style= remains (one mode).

Release Preparation: Operational redeploy only: rebuilt the dev docker image and restarted the plgg-example container on :3001; the shared cloudflared connector and the plgg-example.qmu.dev route were untouched. API surface change: removed the inline style_ and the css name; the single public builder is style_ (plgg-view/style). Internal-only; downstream code uses style_. attr('style', ...) is the documented escape for dynamic inline values.
