+++ packages/plgg-view/src/Program/usecase/render.spec.ts
@@ -23,11 +23,10 @@ import {
   type Motion,
 } from "plgg-view/Html/model/Attribute";
 import { Html } from "plgg-view/Html/model/Html";
-import { style_ } from "plgg-view/Style/usecase/style_";
 import {
-  css,
+  style_,
   hashClass,
-} from "plgg-view/Style/usecase/css";
+} from "plgg-view/Style/usecase/style_";
 import {
   p,
   bg,
@@ -477,35 +476,20 @@ test("tolerates externally-added child nodes when growing a list", () => {
   ).toBe(true);
 });
 
-test("an inline-style utility sets and updates the style attribute", () => {
+test("style_ sets and patches the class attribute (hook + atoms)", () => {
   const root = document.createElement("div");
   const render = makeRenderer<never>(root, noop);
-  render(el("div", [style_(p(2))], []));
+  render(el("div", [style_("box", p(2))], []));
   const node = root.firstElementChild;
-  expect(node?.getAttribute("style")).toBe(
-    "padding:0.5rem",
+  expect(node?.getAttribute("class")).toBe(
+    `box ${hashClass("|padding:0.5rem")}`,
   );
   render(
     el("div", [style_(p(4), bg("primary"))], []),
   );
   expect(root.firstElementChild).toBe(node);
-  expect(node?.getAttribute("style")).toBe(
-    "padding:1rem;background-color:#2563eb",
-  );
-});
-
-test("a css() directive sets and patches the class attribute", () => {
-  const root = document.createElement("div");
-  const render = makeRenderer<never>(root, noop);
-  render(el("div", [css("box", p(2))], []));
-  const node = root.firstElementChild;
   expect(node?.getAttribute("class")).toBe(
-    `box ${hashClass("|padding:0.5rem")}`,
-  );
-  render(el("div", [css(p(4))], []));
-  expect(root.firstElementChild).toBe(node);
-  expect(node?.getAttribute("class")).toBe(
-    hashClass("|padding:1rem"),
+    `${hashClass("|padding:1rem")} ${hashClass("|background-color:#2563eb")}`,
   );
 });
