--- packages/plgg-view/src/Style/usecase/css.spec.ts
@@ -1,75 +1,0 @@
-import { test, expect } from "vitest";
-import {
-  p,
-  bg,
-  color,
-} from "plgg-view/Style/usecase/utilities";
-import {
-  css,
-  hover,
-  focus,
-  active,
-  hashClass,
-} from "plgg-view/Style/usecase/css";
-
-test("hashClass is deterministic and content-addressed", () => {
-  expect(hashClass("x")).toBe(hashClass("x"));
-  expect(hashClass("x")).not.toBe(hashClass("y"));
-  expect(hashClass("x").startsWith("c")).toBe(
-    true,
-  );
-});
-
-test("css(styles) builds one atomic class + rule per declaration", () => {
-  const a = css(p(3));
-  expect(a.__tag).toBe("Css");
-  if (a.__tag === "Css") {
-    const cls = hashClass("|padding:0.75rem");
-    expect(a.content.rules).toEqual([
-      {
-        className: cls,
-        selector: "",
-        prop: "padding",
-        value: "0.75rem",
-      },
-    ]);
-    expect(a.content.classes).toBe(cls);
-  }
-});
-
-test("css prepends literal class hooks (which carry no rule)", () => {
-  const a = css("todo", p(3));
-  if (a.__tag === "Css") {
-    expect(a.content.classes).toBe(
-      "todo " + hashClass("|padding:0.75rem"),
-    );
-    expect(a.content.rules).toHaveLength(1);
-  }
-});
-
-test("variants carry a pseudo-class selector on every atom", () => {
-  const a = css(
-    hover(bg("primary"), color("primary-text")),
-  );
-  if (a.__tag === "Css") {
-    expect(
-      a.content.rules.map((r) => r.selector),
-    ).toEqual([":hover", ":hover"]);
-    expect(
-      a.content.rules.map((r) => r.prop),
-    ).toEqual(["background-color", "color"]);
-    expect(
-      a.content.classes.split(" "),
-    ).toHaveLength(2);
-  }
-  expect(focus(p(1)).selector).toBe(":focus");
-  expect(active(p(1)).selector).toBe(":active");
-});
-
-test("css() with no parts is an empty class set", () => {
-  const a = css();
-  if (a.__tag === "Css") {
-    expect(a.content.classes).toBe("");
-    expect(a.content.rules).toEqual([]);
-  }
-});
