+++ packages/plgg-view/src/Style/usecase/index.ts
@@ -1,3 +1,2 @@
 export * from "plgg-view/Style/usecase/utilities";
 export * from "plgg-view/Style/usecase/style_";
-export * from "plgg-view/Style/usecase/css";
